+++ pkg/services/accesscontrol/ossaccesscontrol/permissions_services.go
@@ -1,4 +1,4 @@
-package resourceservices
+package ossaccesscontrol
 
 import (
 	"context"
@@ -12,26 +12,27 @@ import (
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 )
 
-func ProvideResourceServices(router routing.RouteRegister, sql *sqlstore.SQLStore, ac accesscontrol.AccessControl, store resourcepermissions.Store) (*ResourceServices, error) {
+func ProvidePermissionsServices(router routing.RouteRegister, sql *sqlstore.SQLStore, ac accesscontrol.AccessControl, store resourcepermissions.Store) (*PermissionsService, error) {
 	teamPermissions, err := ProvideTeamPermissions(router, sql, ac, store)
 	if err != nil {
 		return nil, err
 	}
 
-	return &ResourceServices{services: map[string]*resourcepermissions.Service{
-		"teams": teamPermissions,
+	return &PermissionsService{services: map[string]accesscontrol.PermissionsService{
+		"teams":      teamPermissions,
+		"datasource": provideEmptyPermissionsService(),
 	}}, nil
 }
 
-type ResourceServices struct {
-	services map[string]*resourcepermissions.Service
+type PermissionsService struct {
+	services map[string]accesscontrol.PermissionsService
 }
 
-func (s *ResourceServices) GetTeamService() accesscontrol.PermissionsService {
+func (s *PermissionsService) GetTeamService() accesscontrol.PermissionsService {
 	return s.services["teams"]
 }
 
-func (s *ResourceServices) GetDataSourceService() accesscontrol.PermissionsService {
+func (s *PermissionsService) GetDataSourceService() accesscontrol.PermissionsService {
 	// not used in oss
 	return nil
 }
@@ -105,4 +106,32 @@ func ProvideTeamPermissions(router routing.RouteRegister, sql *sqlstore.SQLStore
 	}
 
 	return resourcepermissions.New(options, router, ac, store, sql)
+}
+
+func provideEmptyPermissionsService() accesscontrol.PermissionsService {
+	return &emptyPermissionsService{}
+}
+
+var _ accesscontrol.PermissionsService = new(emptyPermissionsService)
+
+type emptyPermissionsService struct{}
+
+func (e emptyPermissionsService) GetPermissions(ctx context.Context, orgID int64, resourceID string) ([]accesscontrol.ResourcePermission, error) {
+	return nil, nil
+}
+
+func (e emptyPermissionsService) SetUserPermission(ctx context.Context, orgID int64, user accesscontrol.User, resourceID, permission string) (*accesscontrol.ResourcePermission, error) {
+	return nil, nil
+}
+
+func (e emptyPermissionsService) SetTeamPermission(ctx context.Context, orgID, teamID int64, resourceID, permission string) (*accesscontrol.ResourcePermission, error) {
+	return nil, nil
+}
+
+func (e emptyPermissionsService) SetBuiltInRolePermission(ctx context.Context, orgID int64, builtInRole string, resourceID string, permission string) (*accesscontrol.ResourcePermission, error) {
+	return nil, nil
+}
+
+func (e emptyPermissionsService) SetPermissions(ctx context.Context, orgID int64, resourceID string, commands ...accesscontrol.SetResourcePermissionCommand) ([]accesscontrol.ResourcePermission, error) {
+	return nil, nil
 }
